refactor(manager): move some logic from gitlabci-include to gitlabci manager
